pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sortsense"
version = "1.0.0"
description = "Smart file categorization and reorganization using OCR and text extraction"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "SortSense Contributors"}
]
maintainers = [
    {name = "SortSense Contributors"}
]
keywords = [
    "file-organizer",
    "ocr",
    "document-classification",
    "file-management",
    "text-extraction",
    "tesseract",
    "automation",
    "productivity"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Desktop Environment :: File Managers",
    "Topic :: Utilities",
    "Topic :: Text Processing",
    "Topic :: Scientific/Engineering :: Image Recognition"
]

dependencies = [
    "tqdm>=4.60.0",
]

[project.optional-dependencies]
full = [
    "python-docx>=0.8.11",
    "openpyxl>=3.0.0",
    "pandas>=1.3.0",
    "Pillow>=8.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "build>=0.10.0",
    "twine>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/sortsense"
Documentation = "https://github.com/yourusername/sortsense#readme"
Repository = "https://github.com/yourusername/sortsense.git"
Issues = "https://github.com/yourusername/sortsense/issues"
Changelog = "https://github.com/yourusername/sortsense/blob/main/CHANGELOG.md"

[project.scripts]
sortsense = "sortsense.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["sortsense*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py38"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
